# portal/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Runner Portal API",
    "description": "API do portal de assessorias esportivas Omni Runner. Gerencia atletas, treinos, financeiro, OmniCoins e administração da plataforma.",
    "version": "1.0.0",
    "contact": {
      "name": "Omni Runner"
    }
  },
  "servers": [
    {
      "url": "https://portal.omnirunner.app",
      "description": "Produção"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/checkout": {
      "post": {
        "tags": ["Financial"],
        "summary": "Criar sessão de checkout",
        "description": "Inicia checkout para compra de produto (OmniCoins/badges). Retorna URL de redirecionamento.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["product_id"],
                "properties": {
                  "product_id": { "type": "string", "description": "ID do produto" },
                  "gateway": { "type": "string", "enum": ["mercadopago", "stripe"], "default": "mercadopago" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checkout_url": { "type": "string" },
                    "purchase_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado ou validação falhou" },
          "401": { "description": "Não autorizado" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
    "/api/billing-portal": {
      "post": {
        "tags": ["Financial"],
        "summary": "Abrir portal de faturamento",
        "description": "Retorna URL do portal Stripe/Mercado Pago para gestão de assinaturas e pagamentos.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "portal_url": { "type": "string" } }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
    "/api/auto-topup": {
      "post": {
        "tags": ["Financial"],
        "summary": "Configurar auto-recarga",
        "description": "Define ou atualiza configurações de recarga automática de créditos.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" },
                  "threshold_tokens": { "type": "integer", "minimum": 1 },
                  "product_id": { "type": "string" },
                  "max_per_month": { "type": "integer", "minimum": 1, "maximum": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Configuração salva" },
          "400": { "description": "Validação falhou ou product_id obrigatório na criação" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
    "/api/branding": {
      "get": {
        "tags": ["Settings"],
        "summary": "Obter branding",
        "description": "Retorna configurações de marca da assessoria (logo, cores).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logo_url": { "type": "string", "nullable": true },
                    "primary_color": { "type": "string" },
                    "sidebar_bg": { "type": "string" },
                    "sidebar_text": { "type": "string" },
                    "accent_color": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Salvar branding",
        "description": "Atualiza configurações de marca da assessoria.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "logo_url": { "type": "string", "format": "uri", "nullable": true },
                  "primary_color": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" },
                  "sidebar_bg": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" },
                  "sidebar_text": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" },
                  "accent_color": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Branding salvo" },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/clearing": {
      "get": {
        "tags": ["Financial"],
        "summary": "Listar liquidações",
        "description": "Lista liquidações (clearing) da assessoria. Query role: creditor, debtor ou both.",
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["creditor", "debtor", "both"] } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "settlements": { "type": "array", "items": { "type": "object" } } }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin/coach" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT do Supabase (access_token da sessão)"
      }
    },
...
  }
}